+++ CHANGELOG.md
@@ -5,3 +5,5 @@ Additions:
 Fixes:
 - Fix 12.1 lua errors
 - Hopefully fix stuck cooldowns
+- Custom icon cooldown and charge display improvements and fixes
+- Fix the edge and swipe for abilities with charges and an active buff